SAT_platform/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
openai==1.52.0
stripe==9.4.0
alembic==1.13.3
python-multipart==0.0.17
